Add `download-file` Next.js internal API route to enable on Vercel
