+++ pages/api/download-file.js
@@ -1,0 +1,3 @@
+import downloadFileHandler from '../../server/download-file';
+
+export default downloadFileHandler;
